--- Casos de Uso de Machine Learning Supervisado/clase-machine-learning2708/templates/caso.html
@@ -1,87 +1,0 @@
-<!doctype html>
-<html lang="es">
-<head>
-  <meta charset="utf-8">
-  <title>{{ caso.problema }}</title>
-  <!-- Bootstrap -->
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet">
-  <!-- Estilos personalizados -->
-  <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
-</head>
-<body class="bg-light">
-
-  <!-- Barra de navegación -->
-  <nav class="navbar navbar-expand-lg navbar-dark bg-primary mb-4">
-    <div class="container-fluid">
-      <a class="navbar-brand" href="{{ url_for('index') }}">ML Supervisado</a>
-      <a class="nav-link text-white" href="{{ url_for('index') }}">Inicio</a>
-    </div>
-  </nav>
-
-  <main class="container">
-    {% if caso.contenido_html %}
-      <!-- Renderiza el bloque narrativo tal cual -->
-      <div class="card mb-4">
-        <div class="card-body">
-          {{ caso.contenido_html | safe }}
-        </div>
-      </div>
-    {% else %}
-      <!-- Formato estándar si no hay contenido narrativo -->
-      <div class="card mb-4">
-        <div class="card-body">
-          <h1 class="card-title h3">{{ caso.problema }}</h1>
-          <p><strong>Sector y organización:</strong> {{ caso.sector }} — {{ caso.organizacion }}</p>
-          <p><strong>Tipo de algoritmo:</strong> {{ caso.algoritmo }}</p>
-          <p><strong>Tarea supervisada:</strong> {{ caso.tarea }}</p>
-          <p><strong>Datos:</strong> {{ caso.datos }}</p>
-          <p><strong>Métrica de evaluación:</strong> {{ caso.metrica }}</p>
-          <p><strong>Resultados y beneficios:</strong> {{ caso.beneficios }}</p>
-        </div>
-      </div>
-
-      {% if caso.por_que_supervisado or caso.features_clave or caso.validacion or caso.manejo_desbalance %}
-      <div class="card mb-4">
-        <div class="card-header">Profundización técnica</div>
-        <div class="card-body">
-          {% if caso.por_que_supervisado %}
-            <p><strong>Por qué es supervisado:</strong> {{ caso.por_que_supervisado }}</p>
-          {% endif %}
-
-          {% if caso.features_clave %}
-            <h5>Características clave</h5>
-            <ul>
-              {% for f in caso.features_clave %}
-                <li>{{ f }}</li>
-              {% endfor %}
-            </ul>
-          {% endif %}
-
-          {% if caso.validacion %}
-            <p><strong>Estrategia de validación:</strong> {{ caso.validacion }}</p>
-          {% endif %}
-
-          {% if caso.manejo_desbalance %}
-            <p><strong>Manejo del desbalance:</strong> {{ caso.manejo_desbalance }}</p>
-          {% endif %}
-        </div>
-      </div>
-      {% endif %}
-
-      <div class="card">
-        <div class="card-header">Referencias en APA</div>
-        <div class="card-body">
-          <ul class="mb-0">
-            {% for ref in caso.referencia %}
-              <li>{{ ref }}</li>
-            {% endfor %}
-          </ul>
-        </div>
-      </div>
-    {% endif %}
-  </main>
-
-  <!-- Bootstrap JS -->
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
-</body>
-</html>
