+++ clase-machine-learning2708/templates/index.html
@@ -1,0 +1,72 @@
+<!doctype html>
+<html lang="es">
+<head>
+  <meta charset="utf-8">
+  <title>Casos de uso ML supervisado</title>
+  <!-- Bootstrap CDN -->
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet">
+  <!-- Tu hoja de estilos personalizada -->
+  <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
+</head>
+<body class="bg-light">
+
+  <!-- Barra de navegación con menú flotante -->
+  <nav class="navbar navbar-expand-lg navbar-dark bg-primary mb-4">
+    <div class="container-fluid">
+      <a class="navbar-brand" href="{{ url_for('index') }}">ML Supervisado</a>
+      <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav">
+        <span class="navbar-toggler-icon"></span>
+      </button>
+      <div class="collapse navbar-collapse" id="navbarNav">
+        <ul class="navbar-nav">
+          <li class="nav-item"><a class="nav-link active" href="{{ url_for('index') }}">Inicio</a></li>
+          <li class="nav-item dropdown">
+            <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown">Casos</a>
+            <ul class="dropdown-menu">
+              {% for c in casos %}
+              <li>
+                <a class="dropdown-item" href="{{ url_for('caso', id=c.id) }}">
+                  {% if c.sector == 'Entretenimiento' %}🎬{% elif c.sector == 'Salud' %}⚕️{% elif c.sector == 'Finanzas' %}💳{% elif c.sector == 'Consumo masivo' %}🏪{% endif %}
+                  {{ c.sector }} — {{ c.problema }}
+                </a>
+              </li>
+              {% endfor %}
+            </ul>
+          </li>
+        </ul>
+      </div>
+    </div>
+  </nav>
+
+  <!-- Contenido principal con tabla -->
+  <main class="container">
+    <h1 class="mb-4">Casos de uso de Machine Learning supervisado</h1>
+    <div class="table-responsive">
+      <table class="table table-bordered table-hover align-middle">
+        <thead class="table-light">
+          <tr>
+            <th>Problema</th>
+            <th>Algoritmo</th>
+            <th>Sector</th>
+            <th>Ver más</th>
+          </tr>
+        </thead>
+        <tbody>
+          {% for c in casos %}
+          <tr>
+            <td>{{ c.problema }}</td>
+            <td>{{ c.algoritmo }}</td>
+            <td>{{ c.sector }}</td>
+            <td><a href="{{ url_for('caso', id=c.id) }}" class="btn btn-sm btn-outline-primary">Detalles</a></td>
+          </tr>
+          {% endfor %}
+        </tbody>
+      </table>
+    </div>
+  </main>
+
+  <!-- Bootstrap JS -->
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
+</body>
+</html>
+
